refactor tensorrt llm
